fix swagger (#761)

+++ pkg/cmd/workspace/util/repository_wizard.go
@@ -108,8 +108,8 @@ func getRepositoryFromWizard(config RepositoryWizardConfig) (*apiclient.GitRepos
 	}
 
 	var namespaceList []apiclient.GitNamespace
-	page := 1
-	perPage := 100
+	page := int32(1)
+	perPage := int32(100)
 
 	for {
 		var namespaces []apiclient.GitNamespace
@@ -124,7 +124,7 @@ func getRepositoryFromWizard(config RepositoryWizardConfig) (*apiclient.GitRepos
 		namespaceList = append(namespaceList, namespaces...)
 
 		// Break for git providers with unsupported pagination OR on reaching exhausted items.
-		if isGitProviderWithUnsupportedPagination(providerId) || len(namespaces) < perPage {
+		if isGitProviderWithUnsupportedPagination(providerId) || int32(len(namespaces)) < perPage {
 			break
 		}
 
@@ -164,7 +164,7 @@ func getRepositoryFromWizard(config RepositoryWizardConfig) (*apiclient.GitRepos
 		providerRepos = append(providerRepos, repos...)
 
 		// Break for git providers with unsupported pagination OR on reaching exhausted items.
-		if isGitProviderWithUnsupportedPagination(providerId) || len(repos) < perPage {
+		if isGitProviderWithUnsupportedPagination(providerId) || int32(len(repos)) < perPage {
 			break
 		}
